# Implement faster memcmp for x86_64