--- a/src/mem/x86_64.rs
+++ b/src/mem/x86_64.rs
@@ -113,6 +113,7 @@ pub unsafe fn compare_bytes(a: *const u8, b: *const u8, n: usize) -> i32 {
         // Just to be sure we're actually working with powers of two...
         let _ = const { 1 - mem::size_of::<T>().count_ones() }; // <= 1
         let _ = const { mem::size_of::<T>().count_ones() - 1 }; // >= 1
+
         // This should be equivalent to division with power-of-two sizes, except the former
         // somehow still leaves a call to panic because ??
         for _ in 0..n >> mem::size_of::<T>().trailing_zeros() {